feat: (수정, 추가, 삭제)비밀번호 매칭 구현

+++ src/main/java/org/example/memo/service/MemoService.java
@@ -26,6 +26,10 @@ public class MemoService {
     public MemoResponseDto save(MemoRequestDto memoRequestDto) {
         //요청 받은 데이터로 객체 생성
         Memo memo = new Memo(memoRequestDto.getTitle(), memoRequestDto.getContent(), memoRequestDto.getName(), memoRequestDto.getPw());
+        if (memoRequestDto.getPw() == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The title is a required value.");
+        }
+
         //DB에 저장
         Memo saveMemo = memoRepository.save(memo);
 
@@ -59,7 +63,6 @@ public class MemoService {
     //단건 조회
     @Transactional(readOnly = true)
     public List<MemoResponseDto> findById(Long id) {
-        //Memo의 정보를 DB에서 찾아서 Memos에 저장
         Memo memo = memoRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 " + id + " 입니다"));
         //Dto 리스트 생성 (빈 리스트)
         List<MemoResponseDto> memoResponseDtos = new ArrayList<>();
@@ -69,30 +72,41 @@ public class MemoService {
     }
 
     //일정 수정
+    @Transactional
     public MemoResponseDto update(Long id, String title, String content, String name, Long pw) {
         //조회
         Memo memo = memoRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 " + id + " 입니다"));
+
+        //PW 검증
+        if (!memo.getPw().equals(pw)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The title is a required value.");
+        }
 
         // 필수값 검증
         if (title == null || name == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The title is a required value.");
         }
+
+        // DB에 저장
         memo.update(title, content, name, pw);
         Memo saveMemo = memoRepository.save(memo);
         return new MemoResponseDto(memo.getId(), memo.getTitle(), memo.getContent(), memo.getName(), memo.getCreateAt(), memo.getUpdateAt());
     }
 
-
-
     //일정 삭제
     @Transactional()
-    public ResponseEntity<Void> deleteMemo(Long id){
+    public ResponseEntity<Void> deleteMemo(Long id, Long pw){
+
+        //조회
+        Memo memo = memoRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 " + id + " 입니다"));
+        //PW 검증
+        if (!memo.getPw().equals(pw)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The title is a required value.");
+        }
 
         memoRepository.deleteById(id);
         return null;
     }
-
-
 }
